src/trip-planner/requirements.txt
langchain
langgraph
langchain-community
langchain-experimental
langchain-google-community
langchain-google-community[places]
langchain_tavily
langchain_groq
langchain_openai
fastapi
python-dotenv
uvicorn
httpx
requests
azure-identity
mlflow
gradio>=4.44.1
